[Refactor] SongDetailView 로 음악 재생 뷰 통합
