+++ Halmap/Features/SongDetail/SongDetailViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  SongDetailViewModel.swift
+//  Halmap
+//
+//  Created by JeonJimin on 10/8/23.
+//
+
+import Foundation
